+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 Use NodeIndex through out the code to consolidate the node
 index operation:
 
-- Define NodeIndexRange, aka LevelRange, as Range<NodeIndex>
-- Use the NodeIndexRange for MerkleProof
-- Use the NodeIndexRange for MerkleTree
+- Define NodeIndexRange, aka LevelRange, as Range<NodeIndex>.
+- Use NodeIndexRange in MerkleProof.
+- Use NodeIndexRange in MerkleTree.
+- Use NodeIndex in MerkleLeavesMut.
 
 # Version 0.0.8